Refactor: [WalletConnect] lazy-load based on feature toggle (#3048)
* Refactor: [WalletConnect] lazy-load based on feature toggle
* Refactor: update safe-apps-sdk to v9 with ESM support (#3049)
* Move files to features/walletconnect
