+++ package.json
@@ -34,6 +34,9 @@
   },
   "resolutions": {
     "@ethereumjs/common": "^3.2.0",
+    "@walletconnect/core": "^2.10.6",
+    "@walletconnect/ethereum-provider": "^2.10.6",
+    "@walletconnect/utils": "^2.10.6",
     "@web3-onboard/trezor/**/protobufjs": "^7.2.4"
   },
   "dependencies": {
@@ -47,7 +50,7 @@
     "@mui/material": "^5.14.20",
     "@mui/x-date-pickers": "^5.0.20",
     "@reduxjs/toolkit": "^1.9.5",
-    "@safe-global/safe-apps-sdk": "^8.1.0",
+    "@safe-global/safe-apps-sdk": "^9.0.0-next.1",
     "@safe-global/safe-core-sdk": "^3.3.5",
     "@safe-global/safe-core-sdk-utils": "^1.7.4",
     "@safe-global/safe-deployments": "1.25.0",
@@ -59,15 +62,15 @@
     "@spindl-xyz/attribution-lite": "^1.4.0",
     "@tkey-mpc/common-types": "^8.2.2",
     "@truffle/hdwallet-provider": "^2.1.4",
-    "@walletconnect/utils": "^2.10.2",
-    "@walletconnect/web3wallet": "^1.9.2",
+    "@walletconnect/utils": "^2.10.6",
+    "@walletconnect/web3wallet": "^1.9.5",
     "@web3-onboard/coinbase": "^2.2.6",
     "@web3-onboard/core": "^2.21.2",
     "@web3-onboard/injected-wallets": "^2.10.7",
     "@web3-onboard/keystone": "^2.3.7",
     "@web3-onboard/ledger": "2.3.2",
     "@web3-onboard/trezor": "^2.4.2",
-    "@web3-onboard/walletconnect": "^2.5.0",
+    "@web3-onboard/walletconnect": "^2.5.2",
     "@web3auth/mpc-core-kit": "^1.1.3",
     "blo": "^1.1.1",
     "bn.js": "^5.2.1",
