+++ src/components/common/Header/index.tsx
@@ -15,10 +15,10 @@ import SafeLogo from '@/public/images/logo.svg'
 import Link from 'next/link'
 import useSafeAddress from '@/hooks/useSafeAddress'
 import BatchIndicator from '@/components/batch/BatchIndicator'
-import WalletConnectUi from '@/components/walletconnect'
+import WalletConnect from '@/features/walletconnect/components'
 import { PushNotificationsBanner } from '@/components/settings/PushNotifications/PushNotificationsBanner'
-import { useCurrentChain } from '@/hooks/useChains'
-import { hasFeature, FEATURES } from '@/utils/chains'
+import { FEATURES } from '@/utils/chains'
+import { useHasFeature } from '@/hooks/useChains'
 
 type HeaderProps = {
   onMenuToggle?: Dispatch<SetStateAction<boolean>>
@@ -30,8 +30,7 @@ const Header = ({ onMenuToggle, onBatchToggle }: HeaderProps): ReactElement => {
   const safeAddress = useSafeAddress()
   const showSafeToken = safeAddress && !!getSafeTokenAddress(chainId)
   const router = useRouter()
-  const chain = useCurrentChain()
-  const enableWc = !!chain && hasFeature(chain, FEATURES.NATIVE_WALLETCONNECT)
+  const enableWc = useHasFeature(FEATURES.NATIVE_WALLETCONNECT)
 
   // Logo link: if on Dashboard, link to Welcome, otherwise to the root (which redirects to either Dashboard or Welcome)
   const logoHref = router.pathname === AppRoutes.home ? AppRoutes.welcome.index : AppRoutes.index
@@ -84,7 +83,7 @@ const Header = ({ onMenuToggle, onBatchToggle }: HeaderProps): ReactElement => {
 
       {enableWc && (
         <div className={classnames(css.element, css.hideMobile)}>
-          <WalletConnectUi />
+          <WalletConnect />
         </div>
       )}
