+++ src/components/safe-apps/AppFrame/useAppCommunicator.ts
@@ -7,7 +7,6 @@ import type {
   ChainInfo as WebCoreChainInfo,
   TransactionDetails,
 } from '@safe-global/safe-gateway-typescript-sdk'
-import type { Permission, PermissionRequest } from '@safe-global/safe-apps-sdk/dist/src/types/permissions'
 import type {
   AddressBookItem,
   BaseTransaction,
@@ -25,8 +24,8 @@ import type {
   ChainInfo,
   SafeBalances,
 } from '@safe-global/safe-apps-sdk'
-import { Methods } from '@safe-global/safe-apps-sdk'
-import { RPC_CALLS } from '@safe-global/safe-apps-sdk/dist/src/eth/constants'
+import { Methods, RPC_CALLS } from '@safe-global/safe-apps-sdk'
+import type { Permission, PermissionRequest } from '@safe-global/safe-apps-sdk/dist/types/types/permissions'
 import type { SafeSettings } from '@safe-global/safe-apps-sdk'
 import AppCommunicator from '@/services/safe-apps/AppCommunicator'
 import { Errors, logError } from '@/services/exceptions'
