+++ src/components/safe-apps/PermissionsPrompt.tsx
@@ -1,5 +1,5 @@
 import type { ReactElement } from 'react'
-import type { PermissionRequest } from '@safe-global/safe-apps-sdk/dist/src/types/permissions'
+import type { PermissionRequest } from '@safe-global/safe-apps-sdk/dist/types/types/permissions'
 import { Button, Dialog, DialogActions, DialogContent, Divider, Typography } from '@mui/material'
 
 import { ModalDialogTitle } from '@/components/common/ModalDialog'
