+++ src/features/walletconnect/components/WcHeaderWidget/WcIcon.tsx
