+++ src/features/walletconnect/components/WcInput/index.tsx
@@ -1,9 +1,9 @@
 import { useCallback, useContext, useEffect, useState } from 'react'
 import { Button, InputAdornment, TextField } from '@mui/material'
-import { WalletConnectContext } from '@/services/walletconnect/WalletConnectContext'
+import { WalletConnectContext } from '@/features/walletconnect/WalletConnectContext'
 import { asError } from '@/services/exceptions/utils'
 import { getClipboard, isClipboardSupported } from '@/utils/clipboard'
-import { isPairingUri } from '@/services/walletconnect/utils'
+import { isPairingUri } from '@/features/walletconnect/services/utils'
 import Track from '@/components/common/Track'
 import { WALLETCONNECT_EVENTS } from '@/services/analytics/events/walletconnect'
 import { trackEvent } from '@/services/analytics'
