+++ src/features/walletconnect/constants.ts
