+++ src/features/walletconnect/services/__tests__/WalletConnectWallet.test.ts
