+++ src/features/walletconnect/services/__tests__/utils.test.ts
